feat: updated ast and parser

+++ software/luna/include/ast.h
@@ -119,7 +119,7 @@ struct ExpressionNode {
 };
 
 struct IntegerLiteralNode {
-  uint16_t value;
+  int64_t value;
 };
 
 struct SymbolLiteralNode {
